+++ src/oops/SOLID/openClosePrinciple/before/taxes/TaxPayable.java
@@ -3,5 +3,5 @@ package oops.SOLID.openClosePrinciple.before.taxes;
 import oops.SOLID.openClosePrinciple.before.employees.Employee;
 
 public interface TaxPayable {
-    public int calculateTax(Employee employee);
+    public double calculateTax(Employee employee);
 }
